+++ src/renderer/src/Model.cpp
@@ -23,14 +23,14 @@ void Model::bind() const
 
 //-----------------------------------------------------------------------------
 
-u32 Model::getVerteciesCount() const noexcept
+u64 Model::getVerteciesCount() const noexcept
 {
 	return m_vertexArray.getVerticesCount();
 }
 
 //-----------------------------------------------------------------------------
 
-u32 Model::getIndeciesCount() const noexcept
+u64 Model::getIndeciesCount() const noexcept
 {
 	return m_vertexArray.getIndicesCount();
 }
@@ -58,10 +58,10 @@ void Model::calculateBoundingBox(const std::vector<Vertex>& _vertices) noexcept
 	const Vertex& closest		= *maxFinder([](const Vertex& _lhs, const Vertex& _rhs) { return _lhs.pos.z < _rhs.pos.z; });
 	const Vertex& farthest		= *maxFinder([](const Vertex& _lhs, const Vertex& _rhs) { return _lhs.pos.z > _rhs.pos.z; });
 
-	m_boundingBox.frontRect.topLeft		= glm::vec3(rightmost.pos.x, highest.pos.y, farthest.pos.z);
-	m_boundingBox.frontRect.bottomRight	= glm::vec3(leftmost.pos.x, lowest.pos.y, farthest.pos.z);
-	m_boundingBox.backRect.topLeft		= glm::vec3(rightmost.pos.x, highest.pos.y, closest.pos.z);
-	m_boundingBox.backRect.bottomRight	= glm::vec3(leftmost.pos.x, lowest.pos.y, closest.pos.z);
+	m_boundingBox.frontRect.topLeft		= glm::vec3(rightmost.pos.x,	highest.pos.y,	farthest.pos.z);
+	m_boundingBox.frontRect.bottomRight	= glm::vec3(leftmost.pos.x,		lowest.pos.y,	farthest.pos.z);
+	m_boundingBox.backRect.topLeft		= glm::vec3(rightmost.pos.x,	highest.pos.y,	closest.pos.z);
+	m_boundingBox.backRect.bottomRight	= glm::vec3(leftmost.pos.x,		lowest.pos.y,	closest.pos.z);
 }
 
 //-----------------------------------------------------------------------------
